backend/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0

# Database
mysqlclient==2.1.1
redis==4.6.0

# HTTP requests
requests==2.31.0

# Machine Learning
numpy==1.24.3
scikit-learn==1.3.0

# Utilities
Werkzeug==2.3.7
click==8.1.8

# Development and testing
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Production
gunicorn==21.2.0
